requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
groq==0.4.0
markdown==3.5
python-wordpress-xmlrpc==2.3
python-dateutil==2.8.2
